Fix and simplify the self.preinstall section (#3258)

+++ packages/xdg_base.rb
@@ -3,31 +3,18 @@ require 'package'
 class Xdg_base < Package
   description 'XDG Base Directory Specification Configuration'
   homepage 'https://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html'
-  version '0.7-4'
+  version '0.7-5'
   source_url 'file:///dev/null'
   source_sha256 'e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855'
 
   def self.preinstall
     # Save any previous configuration, if it exists.
-    if Dir.exists? "#{HOME}/.config"
-      if File.readlink("#{HOME}/.config") != "#{CREW_PREFIX}/.config"
-        puts('File.readlink("#{HOME}/.config") != "#{CREW_PREFIX}/.config"') # Debug message
-        if !FileUtils.cp_r("#{HOME}/.config", "#{CREW_PREFIX}/")
-          puts('FileUtils.cp_r("#{HOME}/.config", "#{CREW_PREFIX}/")') # Debug
-          FileUtils.rm_rf("#{HOME}/.config")
-        end
-      end
+    if File.directory?("#{HOME}/.config") && !File.symlink?("#{HOME}/.config")
+      FileUtils.rm_rf("#{HOME}/.config") unless FileUtils.cp_r("#{HOME}/.config", "#{CREW_PREFIX}/")
     end
-    if Dir.exists? "#{HOME}/.local"
-      puts('Dir.exists? "#{HOME}/.local"') # Debug
+    if File.directory?("#{HOME}/.local") && !File.symlink?("#{HOME}/.local")
       FileUtils.mkdir_p("#{CREW_PREFIX}/.config") unless Dir.exists? "#{CREW_PREFIX}/.config"
-      if File.readlink("#{HOME}/.local") != "#{CREW_PREFIX}/.config"
-        puts('File.readlink("#{HOME}/.local") != "#{CREW_PREFIX}/.config"') # Debug
-        if !FileUtils.cp_r("#{HOME}/.local/.", "#{CREW_PREFIX}/.config/")
-          puts('FileUtils.cp_r("#{HOME}/.local/.", "#{CREW_PREFIX}/.config/")') # Debug
-          FileUtils.rm_rf("#{HOME}/.local") 
-        end
-      end
+      FileUtils.rm_rf("#{HOME}/.local") unless FileUtils.cp_r("#{HOME}/.local/.", "#{CREW_PREFIX}/.config/")
     end
   end
